requirements.txt
hydra-core
torch
torchvision
numpy
matplotlib
tqdm
wandb
